requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.2
bcrypt==3.2.2 